+++ src/extension.ts
@@ -1,13 +1,12 @@
-import { Marp, MarpOptions } from '@marp-team/marp-core'
+import { Marp } from '@marp-team/marp-core'
 import { ExtensionContext, commands, workspace } from 'vscode'
+import exportCommand from './commands/export' // tslint:disable-line: import-name
 import lineNumber from './plugins/line-number'
 import outline from './plugins/outline'
-
-let cachedMarpOption: MarpOptions | undefined
+import { marpCoreOptionForPreview, clearMarpCoreOptionCache } from './option'
 
 const frontMatterRegex = /^-{3,}\s*([^]*?)^\s*-{3}/m
 const marpDirectiveRegex = /^marp\s*:\s*true\s*$/m
-const marpConfiguration = () => workspace.getConfiguration('markdown.marp')
 const marpVscode = Symbol('marp-vscode')
 const shouldRefreshConfs = [
   'markdown.marp.breaks',
@@ -21,31 +20,6 @@ const detectMarpFromFrontMatter = (markdown: string): boolean => {
   return !!(m && m.index === 0 && marpDirectiveRegex.exec(m[0].trim()))
 }
 
-const marpOption = (baseOpts): MarpOptions => {
-  if (!cachedMarpOption) {
-    const breaks: boolean = (() => {
-      switch (marpConfiguration().get<string>('breaks')) {
-        case 'off':
-          return false
-        case 'inherit':
-          return baseOpts.breaks
-        default:
-          return true
-      }
-    })()
-
-    const zoom =
-      workspace.getConfiguration('window').get<number>('zoomLevel') || 0
-
-    cachedMarpOption = {
-      container: { tag: 'div', id: 'marp-vscode', 'data-zoom': 1.2 ** zoom },
-      html: marpConfiguration().get<boolean>('enableHtml') || undefined,
-      markdown: { breaks },
-    }
-  }
-  return cachedMarpOption
-}
-
 export function extendMarkdownIt(md: any) {
   const { parse, renderer } = md
   const { render } = renderer
@@ -53,7 +27,7 @@ export function extendMarkdownIt(md: any) {
   md.parse = (markdown: string, env: any) => {
     // Generate tokens by Marp if enabled
     if (detectMarpFromFrontMatter(markdown)) {
-      md[marpVscode] = new Marp(marpOption(md.options))
+      md[marpVscode] = new Marp(marpCoreOptionForPreview(md.options))
         .use(outline)
         .use(lineNumber)
 
@@ -94,9 +68,13 @@ export function extendMarkdownIt(md: any) {
 
 export const activate = ({ subscriptions }: ExtensionContext) => {
   subscriptions.push(
+    commands.registerCommand('markdown.marp.export', exportCommand)
+  )
+
+  subscriptions.push(
     workspace.onDidChangeConfiguration(e => {
       if (shouldRefreshConfs.some(c => e.affectsConfiguration(c))) {
-        cachedMarpOption = undefined
+        clearMarpCoreOptionCache()
         commands.executeCommand('markdown.preview.refresh')
       }
     })